Move errors and helpers generators to separate files
This patch moves the errors and helpers generator from the `clients.go`
file to new `errors.go` and `helpers.go` files, because the generated
code is used by both clients and servers.
