+++ go.mod
@@ -5,12 +5,10 @@ go 1.12
 require (
 	github.com/antlr/antlr4 v0.0.0-20181218183524-be58ebffde8e
 	github.com/golang/glog v0.0.0-20160126235308-23def4e6c14b
-	github.com/golang/protobuf v1.3.1 // indirect
 	github.com/gorilla/mux v1.7.3
 	github.com/onsi/ginkgo v1.9.0
 	github.com/onsi/gomega v1.5.0
+	github.com/openshift-online/ocm-sdk-go v0.1.43
 	github.com/spf13/cobra v0.0.5
 	github.com/spf13/pflag v1.0.3
-	golang.org/x/net v0.0.0-20190322120337-addf6b3196f6 // indirect
-	golang.org/x/sys v0.0.0-20190322080309-f49334f85ddc // indirect
 )
