+++ pkg/generators/clients.go
@@ -137,16 +137,6 @@ func (b *ClientsGeneratorBuilder) Build() (generator *ClientsGenerator, err erro
 func (g *ClientsGenerator) Run() error {
 	var err error
 
-	// Generate the common code:
-	err = g.generateErrors()
-	if err != nil {
-		return err
-	}
-	err = g.generateHelpers()
-	if err != nil {
-		return err
-	}
-
 	// Generate the client for each service:
 	for _, service := range g.model.Services() {
 		g.reporter.Infof("Generating client for service '%s'", service.Name())
@@ -167,400 +157,6 @@ func (g *ClientsGenerator) Run() error {
 	}
 
 	return nil
-}
-
-func (g *ClientsGenerator) generateErrors() error {
-	var err error
-
-	// Calculate the package and file name:
-	pkgName := g.errorsPkg()
-	fileName := g.errorsFile()
-
-	// Create the buffer for the generated code:
-	g.buffer, err = golang.NewBufferBuilder().
-		Reporter(g.reporter).
-		Output(g.output).
-		Base(g.base).
-		Package(pkgName).
-		File(fileName).
-		Build()
-	if err != nil {
-		return err
-	}
-
-	// Generate the code:
-	g.buffer.Import(path.Join(g.base, g.helpersPkg()), "")
-	g.buffer.Import("github.com/golang/glog", "")
-	g.buffer.Emit(`
-		// Error kind is the name of the type used to represent errors.
-		const ErrorKind = "Error"
-
-		// ErrorNilKind is the name of the type used to nil errors.
-		const ErrorNilKind = "ErrorNil"
-
-		// ErrorBuilder is a builder for the error type.
-		type ErrorBuilder struct{
-			id     *string
-			href   *string
-			code   *string
-			reason *string
-		}
-
-		// Error represents errors.
-		type Error struct {
-			id     *string
-			href   *string
-			code   *string
-			reason *string
-		}
-
-		// NewError returns a new ErrorBuilder
-		func NewError() *ErrorBuilder {
-			return new(ErrorBuilder)
-		}
-
-		// ID sets the id field for the ErrorBuilder
-		func (e *ErrorBuilder) ID(id string) *ErrorBuilder {
-			e.id = &id
-			return e
-		}
-
-		// HREF sets the href field for the ErrorBuilder
-		func (e *ErrorBuilder) HREF(href string) *ErrorBuilder {
-			e.href = &href
-			return e
-		}
-
-		// Code sets the cpde field for the ErrorBuilder
-		func (e *ErrorBuilder) Code(code string) *ErrorBuilder {
-			e.code = &code
-			return e
-		}
-
-		// Reason sets the reason field for the ErrorBuilder
-		func (e *ErrorBuilder) Reason(reason string) *ErrorBuilder {
-			e.reason = &reason
-			return e
-		}
-
-		// Build builds a new error type or returns an error.
-		func (e *ErrorBuilder) Build() (*Error, error) {
-			err := new(Error)
-			err.reason = e.reason
-			err.code = e.code
-			err.id = e.id
-			err.href = e.href
-			return err, nil
-		}
-
-		// Kind returns the name of the type of the error.
-		func (e *Error) Kind() string {
-			if e == nil {
-				return ErrorNilKind
-			}
-			return ErrorKind
-		}
-
-		// ID returns the identifier of the error.
-		func (e *Error) ID() string {
-			if e != nil && e.id != nil {
-				return *e.id
-			}
-			return ""
-		}
-
-		// GetID returns the identifier of the error and a flag indicating if the
-		// identifier has a value.
-		func (e *Error) GetID() (value string, ok bool) {
-			ok = e != nil && e.id != nil
-			if ok {
-				value = *e.id
-			}
-			return
-		}
-
-		// HREF returns the link to the error.
-		func (e *Error) HREF() string {
-			if e != nil && e.href != nil {
-				return *e.href
-			}
-			return ""
-		}
-
-		// GetHREF returns the link of the error and a flag indicating if the
-		// link has a value.
-		func (e *Error) GetHREF() (value string, ok bool) {
-			ok = e != nil && e.href != nil
-			if ok {
-				value = *e.href
-			}
-			return
-		}
-
-		// Code returns the code of the error.
-		func (e *Error) Code() string {
-			if e != nil && e.code != nil {
-				return *e.code
-			}
-			return ""
-		}
-
-		// GetCode returns the link of the error and a flag indicating if the
-		// code has a value.
-		func (e *Error) GetCode() (value string, ok bool) {
-			ok = e != nil && e.code != nil
-			if ok {
-				value = *e.code
-			}
-			return
-		}
-
-		// Reason returns the reason of the error.
-		func (e *Error) Reason() string {
-			if e != nil && e.reason != nil {
-				return *e.reason
-			}
-			return ""
-		}
-
-		// GetReason returns the link of the error and a flag indicating if the
-		// reason has a value.
-		func (e *Error) GetReason() (value string, ok bool) {
-			ok = e != nil && e.reason != nil
-			if ok {
-				value = *e.reason
-			}
-			return
-		}
-
-		// Error is the implementation of the error interface.
-		func (e *Error) Error() string {
-			if e.reason != nil {
-				return *e.reason
-			}
-			if e.code != nil {
-				return *e.code
-			}
-			if e.id != nil {
-				return *e.id
-			}
-			return "unknown error"
-		}
-
-		// UnmarshalError reads an error from the given which can be an slice of bytes, a
-		// string, a reader or a JSON decoder.
-		func UnmarshalError(source interface{}) (object *Error, err error) {
-			decoder, err := helpers.NewDecoder(source)
-			if err != nil {
-				return
-			}
-			data := new(errorData)
-			err = decoder.Decode(data)
-			if err != nil {
-				return
-			}
-			object, err = data.unwrap()
-			return
-		}
-
-		// MarshalError writes an error to the given destination which can be an slice of bytes, a
-		// string, a reader or a JSON decoder.
-		func (e *Error) MarshalError(destination interface{}) error {
-			encoder, err := helpers.NewEncoder(destination)
-			if err != nil {
-				return err
-			}
-			object, err := e.wrap()
-			if err != nil {
-				return err
-			}
-			err = encoder.Encode(object)
-			if err != nil {
-				return err
-			}
-			return nil
-		}
-
-		// errorData is the data structure used internally to marshal and unmarshal errors.
-		type errorData struct {
-			Kind   *string "json:\"kind,omitempty\""
-			ID     *string "json:\"id,omitempty\""
-			HREF   *string "json:\"href,omitempty\""
-			Code   *string "json:\"code,omitempty\""
-			Reason *string "json:\"reason,omitempty\""
-		}
-
-		// unwrap is the method used internally to convert the JSON unmarshalled data to an
-		// error.
-		func (d *errorData) unwrap() (object *Error, err error) {
-			if d == nil {
-				return
-			}
-			object = new(Error)
-			if d.Kind != nil && *d.Kind != ErrorKind {
-				err = fmt.Errorf(
-					"expected kind '%s' but got '%s'",
-					ErrorKind, *d.Kind,
-				)
-				return
-			}
-			object.id = d.ID
-			object.href = d.HREF
-			object.code = d.Code
-			object.reason = d.Reason
-			return
-		}
-
-		// wrap is the method used internally to convert the JSON unmarshalled data to an
-		// error.
-		func (d *Error) wrap() (object *errorData, err error) {
-			if d == nil {
-				return
-			}
-			object = new(errorData)
-			if d.Kind() != "" && d.Kind() != ErrorKind {
-				err = fmt.Errorf(
-					"expected kind '%s' but got '%s'",
-					ErrorKind, d.Kind(),
-				)
-				return
-			}
-			object.ID = d.id
-			object.HREF = d.href
-			object.Code = d.code
-			object.Reason = d.reason
-			return
-		}
-
-		var panicID = "1000"
-		var panicError, _ = NewError().
-			ID(panicID).
-			Reason("An unexpected error happened, please check the log of the service " +
-			"for details").
-			Build()
-
-		// SendError writes a given error and status code to a response writer.
-		// if an error occured it will log the error and exit.
-		// This methods is used internaly and no backwards compatibily is guaranteed.
-		func SendError(w http.ResponseWriter, r *http.Request, error *Error) {
-			w.Header().Set("Content-Type", "application/json")
-
-			status, err := strconv.Atoi(error.ID())
-			if err != nil {
-				SendPanic(w, r)
-				return
-			}
-
-			w.WriteHeader(status)
-			err = error.MarshalError(w)
-			if err != nil {
-				glog.Errorf("Can't send response body for request '%s'", r.URL.Path)
-				return
-			}
-		}
-
-		// SendPanic sends a panic error response to the client, but it doesn't end the process.
-		// This methods is used internaly and no backwards compatibily is guaranteed.
-		func SendPanic(w http.ResponseWriter, r *http.Request) {
-			w.Header().Set("Content-Type", "application/json")
-			// Convert it to JSON:
-			err := panicError.MarshalError(w)
-			if err != nil {
-				glog.Errorf(
-					"Can't send panic response for request '%s': %s",
-					r.URL.Path,
-					err.Error(),
-				)
-			}
-		}
-        `)
-
-	// Write the generated code:
-	return g.buffer.Write()
-}
-
-func (g *ClientsGenerator) generateHelpers() error {
-	var err error
-
-	// Calculate the package and file name:
-	pkgName := g.helpersPkg()
-	fileName := g.clientsFile()
-
-	// Create the buffer for the generated code:
-	g.buffer, err = golang.NewBufferBuilder().
-		Reporter(g.reporter).
-		Output(g.output).
-		Base(g.base).
-		Package(pkgName).
-		File(fileName).
-		Build()
-	if err != nil {
-		return err
-	}
-
-	// Generate the code:
-	g.buffer.Import("fmt", "")
-	g.buffer.Import("net/http", "")
-	g.buffer.Import("net/url", "")
-	g.buffer.Import("time", "")
-	g.buffer.Emit(`
-		// AddValue creates the given set of query parameters if needed, an then adds
-		// the given parameter.
-		func AddValue(query *url.Values, name string, value interface{}) {
-			if *query == nil {
-				*query = make(url.Values)
-			}
-			query.Add(name, fmt.Sprintf("%v", value))
-		}
-
-		// CopyQuery creates a copy of the given set of query parameters.
-		func CopyQuery(query url.Values) url.Values {
-			if query == nil {
-				return nil
-			}
-			result := make(url.Values)
-			for name, values := range query {
-				result[name] = CopyValues(values)
-			}
-			return result
-		}
-
-		// AddHeader creates the given set of headers if needed, and then adds the given
-		// header:
-		func AddHeader(header *http.Header, name string, value interface{}) {
-			if *header == nil {
-				*header = make(http.Header)
-			}
-			header.Add(name, fmt.Sprintf("%v", value))
-		}
-
-		// SetHeader creates a copy of the given set of headers, and adds the header
-		// containing the given metrics path.
-		func SetHeader(header http.Header, metric string) http.Header {
-			result := make(http.Header)
-			for name, values := range header {
-				result[name] = CopyValues(values)
-			}
-			result.Set(metricHeader, metric)
-			return result
-		}
-
-		// CopyValues copies a slice of strings.
-		func CopyValues(values []string) []string {
-			if values == nil {
-				return nil
-			}
-			result := make([]string, len(values))
-			copy(result, values)
-			return result
-		}
-
-		// Name of the header used to contain the metrics path:
-		const metricHeader = "X-Metric"
-        `)
-
-	// Write the generated code:
-	return g.buffer.Write()
 }
 
 func (g *ClientsGenerator) generateServiceClient(service *concepts.Service) error {
@@ -598,10 +194,6 @@ func (g *ClientsGenerator) generateServiceClient(service *concepts.Service) erro
 
 	// Generate the clients for the versions:
 	for _, version := range service.Versions() {
-		err = g.generateVersionErrors(version)
-		if err != nil {
-			return err
-		}
 		err = g.generateVersionClient(version)
 		if err != nil {
 			return err
@@ -654,52 +246,6 @@ func (g *ClientsGenerator) generateServiceClientSource(service *concepts.Service
 		"Service", service,
 	)
 
-	return nil
-}
-
-func (g *ClientsGenerator) generateVersionErrors(version *concepts.Version) error {
-	var err error
-
-	// Calculate the package and file name:
-	pkgName := g.pkgName(version)
-	fileName := g.errorsFile()
-
-	// Create the buffer for the generated code:
-	g.buffer, err = golang.NewBufferBuilder().
-		Reporter(g.reporter).
-		Output(g.output).
-		Base(g.base).
-		Package(pkgName).
-		File(fileName).
-		Function("errorName", g.errorName).
-		Build()
-	if err != nil {
-		return err
-	}
-
-	// Generate the code:
-	err = g.generateVersionErrorsSource(version)
-	if err != nil {
-		return err
-	}
-
-	// Write the generated code:
-	return g.buffer.Write()
-}
-
-func (g *ClientsGenerator) generateVersionErrorsSource(version *concepts.Version) error {
-	g.buffer.Emit(`
-		{{ if .Version.Errors }}
-			const (
-				{{ range .Version.Errors }}
-					{{ lineComment .Doc }}
-					{{ errorName . }} = {{ .Code }}
-				{{ end }}
-			)
-		{{ end }}
-		`,
-		"Version", version,
-	)
 	return nil
 }
 
@@ -1164,16 +710,8 @@ func (g *ClientsGenerator) helpersPkg() string {
 	return g.names.Package(nomenclator.Helpers)
 }
 
-func (g *ClientsGenerator) errorsFile() string {
-	return g.names.File(nomenclator.Errors)
-}
-
 func (g *ClientsGenerator) clientsFile() string {
 	return g.names.File(nomenclator.Clients)
-}
-
-func (g *ClientsGenerator) errorName(err *concepts.Error) string {
-	return g.names.Public(names.Cat(err.Name(), nomenclator.Error))
 }
 
 func (g *ClientsGenerator) versionName(version *concepts.Version) string {
